+++ src/Zen.Cli/CommandsOld/Misc/GenerateAdviseCommand.cs
@@ -1,28 +1,29 @@
+using System.ComponentModel;
 using System.Threading.Tasks;
 using Spectre.Console;
 using Spectre.Console.Cli;
 using Zen.Core.Services;
 using static Spectre.Console.AnsiConsole;
 
-namespace Zen.CLI.Commands.Misc
+namespace Zen.Cli.Commands.Misc
 {
-    public class GenerateInsultCommand : AsyncCommand
+    public class GenerateAdviseCommand : AsyncCommand
     {
         private readonly MiscApiService miscApi;
 
-        public GenerateInsultCommand(MiscApiService miscApi)
+        public GenerateAdviseCommand(MiscApiService miscApi)
         {
             this.miscApi = miscApi;
         }
         public override async Task<int> ExecuteAsync(CommandContext context)
         {
-            var insult = await Status()
+            var advise = await Status()
                 .Spinner(Spinner.Known.Dots)
                 .StartAsync("Generating...", async ctx =>
                 {
-                    return await miscApi.GenerateInsultAsync();
+                    return await miscApi.GenerateAdviseAsync();
                 });
-            WriteLine(insult);
+            WriteLine(advise);
             return 0;
         }
     }