+++ src/Zen.Cli/CommandsOld/Misc/GenerateInsultCommand.cs
@@ -1,29 +1,28 @@
-using System.ComponentModel;
 using System.Threading.Tasks;
 using Spectre.Console;
 using Spectre.Console.Cli;
 using Zen.Core.Services;
 using static Spectre.Console.AnsiConsole;
 
-namespace Zen.CLI.Commands.Misc
+namespace Zen.Cli.Commands.Misc
 {
-    public class GenerateAdviseCommand : AsyncCommand
+    public class GenerateInsultCommand : AsyncCommand
     {
         private readonly MiscApiService miscApi;
 
-        public GenerateAdviseCommand(MiscApiService miscApi)
+        public GenerateInsultCommand(MiscApiService miscApi)
         {
             this.miscApi = miscApi;
         }
         public override async Task<int> ExecuteAsync(CommandContext context)
         {
-            var advise = await Status()
+            var insult = await Status()
                 .Spinner(Spinner.Known.Dots)
                 .StartAsync("Generating...", async ctx =>
                 {
-                    return await miscApi.GenerateAdviseAsync();
+                    return await miscApi.GenerateInsultAsync();
                 });
-            WriteLine(advise);
+            WriteLine(insult);
             return 0;
         }
     }