+++ src/Zen.Cli/Program.cs
@@ -1,0 +1,8 @@
+﻿using System.Threading.Tasks;
+using Zen.Core.SpectreConsole;
+using Zen.Cli;
+
+await SpectreConsoleHost
+    .WithStartup<Startup>(args)
+    .UseConfigurator<Startup>()
+    .RunAsync(args);